refactor: Move L7Log related code to separate crate to avoid windows compilation error
