+++ agent/crates/l7-log-derive/tests/l7_protocol/missing_fields.stderr
