+++ agent/crates/public/Cargo.toml
@@ -22,12 +22,10 @@ pnet = "^0.29"
 prost.workspace = true
 rand = "0.8.5"
 regex.workspace = true
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0.72"
-serde_yaml = "0.8"
+serde.workspace = true
+serde_json.workspace = true
+serde_yaml.workspace = true
 socket2.workspace = true
-strum = { version = "0.27", features = ["derive"] }
-strum_macros = "0.27"
 thiserror = "1.0"
 tonic.workspace = true
